Changed the name that displays in Mod Menu and.
Changed the name that displays in Mod Menu, because all other mods start with capital letters, and this one looked ugly, also, for some reason either the lowercase letters or the hyphen put it at the bottom of the list, therefore putting it below World Edit etc, so it broke alphabetical order. I tested on current version by changing the jar file with 7zip, and it displays much more nicely in Mod Menu. I don't want to have to edit this every time for every version, so I am proposing this change.